Added redux to project and reorganized and updated directory structure
