+++ package.json
@@ -39,6 +39,7 @@
     "react-fontawesome": "^1.6.1",
     "react-ga": "2.2.0",
     "react-minimap": "0.7.2",
+    "react-redux": "^5.0.6",
     "react-router-dom": "4.2.2",
     "react-scripts": "0.9.5",
     "react-tap-event-plugin": "2.0.1",
