+++ src/apps/AppManager.jsx
@@ -1,6 +1,6 @@
-import FullApp from './FullApp';
-import SlimApp from './SlimApp';
-import LandingPage from './LandingPage';
+import FullApp from '../containers/FullApp';
+import SlimApp from '../containers/SlimApp';
+import LandingPage from '../components/LandingPage';
 
 class AppManager {
     constructor() {
@@ -17,32 +17,32 @@ class AppManager {
                 app: SlimApp,
                 isExact: true,
                 shortcutConfigurations: {
-                    'Disease Status': { 
-                        referenceDateEnabled: false 
+                    'Disease Status': {
+                        referenceDateEnabled: false
                     },
                     'Toxicity': {   
                         gradesToDisplay: [3,4,5],
                         gradesPrompt: ' PATINA only calculates its endpoint based on adverse events of grades 2 through 5; therefore, only those are shown below. ',
                         topAdverseEvents: [
-                            'Febrile neutropenia', 
-                            'Fatigue', 
-                            'Generalized muscle weakness', 
-                            'Anemia', 
-                            'Diarrhea', 
-                            'Nausea', 
-                            'Platelet count decreased', 
-                            'Anorexia', 
-                            'Constipation', 
-                            'Bullous dermatitis', 
-                            'Upper respiratory infection', 
-                            'Vomiting', 
-                            'Hot flashes', 
-                            'Arthralgia', 
-                            'Osteoporosis', 
-                            'Fever', 
-                            'Infusion related reaction', 
+                            'Febrile neutropenia',
+                            'Fatigue',
+                            'Generalized muscle weakness',
+                            'Anemia',
+                            'Diarrhea',
+                            'Nausea',
+                            'Platelet count decreased',
+                            'Anorexia',
+                            'Constipation',
+                            'Bullous dermatitis',
+                            'Upper respiratory infection',
+                            'Vomiting',
+                            'Hot flashes',
+                            'Arthralgia',
+                            'Osteoporosis',
+                            'Fever',
+                            'Infusion related reaction',
                             'Headache',
-                            'Cough', 
+                            'Cough',
                             'Dyspnea'
                         ]
                     }
@@ -55,10 +55,10 @@ class AppManager {
                 app: SlimApp,
                 isExact: true,
                 shortcutConfigurations: {
-                    'Disease Status': { 
-                        referenceDateEnabled: false 
+                    'Disease Status': {
+                        referenceDateEnabled: false
                     },
-                    'Toxicity': {   
+                    'Toxicity': {
                         gradesToDisplay: [2,3,4,5],
                         gradesPrompt: '  only calculates its endpoint based on adverse events of grades 2 through 5; therefore, only those are shown below. ',
                         topAdverseEvents: [
@@ -82,7 +82,7 @@ class AppManager {
                         ]
                     }
                 }
-            },            
+            },
             {
                 path: '/patient',
                 display: 'Flux Notes',
@@ -100,7 +100,7 @@ class AppManager {
             }
         ];
     }
-    
+
     getSupportedApps() {
         return this.apps;
     }
